Modularize attachment backup validation
Separate request identity, ledger response, timestamp, and size validation from attachment backup route orchestration without changing protocol behavior.
